Add files via upload
Fixed bugs 
A) an error annotating subject names on the outliers   I resolved this by commenting out lines 46-48  in plotter.py because I couldn’t figure it out
B) Error that occurs if subject list or atlas list has only 1 line in the txt file.  
C) Error with the heatmaps. "ValueError: minvalue must be less than or equal to maxvalue”